fix(notmuch): query notmuch if file is missing
During `nm_mbox_sync()` the call to `mh_sync_mailbox_message()` can fail
if it's unable to rename the file. Failure to rename may occur if
underlying file is missing.

Fortunately, if the file was moved by notmuch (not the integration), it
knows the new location. In this case, query notmuch for an updated
filepath and retry the sync. If it fails this time then there's no way
to recover. We don't know where the file went and neither does our
source of truth (notmuch.)